requirement_dev.txt
Django==4.0.4
django-environ==0.11.2
djangorestframework==3.15.1
openpyxl==3.1.2
psycopg2==2.9.3